// src/documents/interfaces/documentsResponse.ts
// Generated by https://quicktype.io

export interface DocumentResponse {
  success: boolean;
  data: Data;
}

export interface Data {
  DOC_ID: number;
  DOC_NOMBRE: string;
  DOC_CODIGO: string;
  DOC_CONTENIDO: string;
  DOC_ID_TIPO: string;
  DOC_ID_PROCESO: number;
  created_at: string;
  updated_at: string;
}